DRTII-1205 version flightMessage for code readibility
